Separating IR cli and thread execution context
